Use the type system better for events
The DomEvent type was not actually a DomEvent but a ThreeEvent, and it wasn't used consistently. The actual DOM events have MouseEvent as a base class (which both WheelEvent and PointerEvent inherit) so I've simplified the code. I've also added a type for the pointer capture functions so we don't need to use `as any` so much.

I've also fixed an error in the type of onPointerMissed. It claimed to take a ThreeEvent<PointerEvent>, but we only ever passed it the raw event (i.e. no intersection stuff) and the raw event could be any MouseEvent. Using DomEvent everywhere, even for actual DOM events, was masking this.

NB This is a breaking change to the API.