Events do not bubble through objects, only up
Previously, events would first bubble through objects i.e. along the ray to farther objects, before then bubbling to the ancestors of those objects. Now, the event only goes to the first object in the path of the ray, and then bubbles to its ancestors. This makes all objects that have at least one event handler "opaque" to mouse/pointer events.

This also removes how hovered objects are tracked. Only the first hit object counts as hovered, but the pointerOver events can bubble to its ancestors until stopPropagation is called. stopPropagation only affects the event it is called on: it does not cause pointerOut events to be delivered to ancestors that had previously had pointerOver bubbled up to them, and it does not stop subsequent pointerOver/pointerMove events.

The pointer capture target, if any, counts as hovered, as in the DOM, but capturing does not cause pointerOver to be sent, and releasing capture does not cause pointerOut to be sent. This may turn out to be wrong but it's not causing a problem for now.

The event tests have been significantly expanded.