set hovered state correctly when capturing with groups
A new test exercises using pointer capture from an ancestor of the event target, and ensures that pointerout still happens after the capture is released. This didn't work so I've improved the code. It now looks at the capturing object as well as the object the event was originally delivered to - though this is a bit academic for now, as we don't differentiate between calling setPointerCapture on target vs. currentTarget. That's something to fix in future, perhaps at the same time as allowing setPointerCapture to be called without an event, like in DOM.