Fix pointer-out/leave events sometimes not called.
When the leaf element of an object tree did not have event handlers, we were failing to call any onPointerOut/onPointerLeave handlers of the parent nodes.

This happened because hover state was being tracked according to the hovered leaf element. However, if the leaf had no event handlers the intersection handling code exited early - skipping the marking of the hover state.

This commit fixes this issue by only allowing an early exit for non-leaf nodes. It also adds a test for this scenario, and removes a second superfluous early exit check.